requirements.txt
Faker==4.1.0
numpy==1.18.4
matplotlib==2.2.3
py-enigma==0.1
tensorflow==2.1.0
tf-nightly==2.3.0.dev2020052101